app/requirements.txt
fastapi==0.85.0
PyJWT==2.4.0
python-decouple==3.6
requests==2.27.1
SQLAlchemy==1.4.41
uvicorn
psycopg2-binary
pydantic[email]
pymongo




